fix(network): dedup ChainSync/LocalChainSync servers, fix StMustReply wedge + intersect cursor poisoning
Fixes #869, #868, #876; implements the #881 dedup tracker.

- #881: extract the ChainSync/LocalChainSync serve loop (cursor tracking,
  MsgFindIntersect/MsgRequestNext handling, rollback propagation, StMustReply
  retry loop) into a shared ServeCore (chainsync::serve_core), parameterized
  over a plain-fn payload encoder (HFC header for N2N, Serialised full block
  for N2C) and a protocol label. ChainSyncServer and LocalChainSyncServer are
  now thin wrappers.

- #869: handle_request_next now loops the post-MsgAwaitReply select! until a
  message is actually sent. Previously any "nothing to serve" outcome (spurious
  announcement wake, Lagged with nothing new, timeout re-poll at a quiescent
  tip) returned Ok(()) without sending anything, silently handing agency back
  while the client — already in StMustReply per the Ouroboros state machine —
  would never send another MsgRequestNext, wedging the connection until the
  client's own timeout.

- #868: porting N2N's logic to N2C via the shared core gives LocalChainSyncServer
  the fixes it was missing: duplicate-serve dedup, biased rollback-first
  select! ordering, the StMustReply retry loop, cursor_at_origin genesis-EBB
  handling (a block at slot 0 is no longer silently skipped), and lagged-
  rollback safety (Bug J cursor revalidation after a fork switch). N2C also
  gains the StMustReply timeout arm it never had.

- #876: MsgFindIntersect now accepts a Point::Specific only when the hash is
  on the canonical chain (is_on_chain, not merely has_block — rejects fork
  blocks) AND the client-claimed slot matches the block's actual on-chain
  slot (find_chain_ancestor). Prevents a client from poisoning the follower
  cursor with an arbitrary (claimed_slot, real_hash) pair, e.g. (u64::MAX,
  real_hash), which would wedge every subsequent lookup.

Test mock BlockProviders (MockBlockProvider, MutableMockBlockProvider,
ForkAwareMockProvider) move to serve_core::test_support so both N2N and N2C
test suites exercise the identical BlockProvider semantics against the
shared core.

New tests:
- chainsync::server: spurious_announcement_wake_does_not_return_agency_without_reply,
  find_intersect_wrong_claimed_slot_rejected, find_intersect_fork_only_hash_rejected,
  find_intersect_correct_canonical_point_found
- local_chainsync::server: fork_switch_sends_rollback_then_serves_new_blocks,
  genesis_block_at_slot_zero_served_from_origin, idle_timeout_repolls_and_eventually_serves
- local_chainsync::server::request_next_at_tip_sends_await_reply updated: it
  previously relied on the #869 bug (a spurious announcement terminating the
  select early) to complete; now asserts the fixed no-reply-on-spurious-wake
  behaviour and aborts the task instead of awaiting a handle that would never
  resolve under the corrected semantics.
