+++ crates/dugite-network/src/protocol/chainsync/server.rs
@@ -28,15 +28,19 @@ use std::time::Duration;
 use minicbor::{Decoder, Encoder};
 use tokio::sync::broadcast;
 
-use crate::codec::Point;
 use crate::error::ProtocolError;
 use crate::mux::channel::MuxChannel;
 use crate::BlockProvider;
 
-use super::{decode_message, encode_message, ChainSyncMessage};
+use super::serve_core::ServeCore;
 
 // Re-use shared HFC helpers from the protocol module.
 use crate::protocol::{storage_era_tag_to_hfc_index, CBOR_TAG_EMBEDDED};
+
+#[cfg(test)]
+use super::{decode_message, encode_message, ChainSyncMessage};
+#[cfg(test)]
+use crate::codec::Point;
 
 /// Extract the block header from raw HFC-wrapped block CBOR and encode it as
 /// `[hfc_index, #6.24(bstr(header_cbor))]` ready for inlining into MsgRollForward.
@@ -150,20 +154,27 @@ pub const CHAINSYNC_MUST_REPLY_TIMEOUT_MIN: Duration = Duration::from_secs(601);
 /// [`CHAINSYNC_MUST_REPLY_TIMEOUT_MIN`]).  Issue #701.
 pub const CHAINSYNC_MUST_REPLY_TIMEOUT_MAX: Duration = Duration::from_secs(911);
 
+/// Encode the `MsgRollForward` payload for N2N ChainSync: extract just the
+/// block header and HFC-wrap it.  Wraps [`extract_header_for_chainsync`]'s
+/// `String` error into a [`ProtocolError::CborDecode`] to match the
+/// `PayloadEncoder` signature used by the shared serve core (issue #881).
+fn n2n_payload_encoder(block_cbor: &[u8]) -> Result<Vec<u8>, ProtocolError> {
+    extract_header_for_chainsync(block_cbor).map_err(|reason| ProtocolError::CborDecode {
+        protocol: "ChainSync",
+        reason: format!("header extraction failed: {reason}"),
+    })
+}
+
 /// ChainSync server that serves headers to a single downstream peer.
+///
+/// A thin wrapper around the shared `ServeCore` (issue #881) — all protocol
+/// logic (cursor tracking, `MsgFindIntersect`/`MsgRequestNext` handling,
+/// rollback propagation, `StMustReply` retry loop) lives in the shared core.
+/// This type only supplies the N2N-specific payload encoder
+/// ([`n2n_payload_encoder`]: HFC-wrapped header, not the full block) and the
+/// `"ChainSync"` protocol label.
 pub struct ChainSyncServer {
-    /// Current cursor: the last point served to this peer.
-    cursor_slot: u64,
-    cursor_hash: [u8; 32],
-    /// Whether the cursor has been initialized (via intersection or genesis).
-    cursor_initialized: bool,
-    /// True when the cursor is at Origin — meaning no block has been served
-    /// yet and we must include blocks at slot 0 (e.g. Byron genesis EBB).
-    cursor_at_origin: bool,
-    /// Per-connection `StMustReply` timeout, drawn once at construction
-    /// uniformly from `[CHAINSYNC_MUST_REPLY_TIMEOUT_MIN, CHAINSYNC_MUST_REPLY_TIMEOUT_MAX]`.
-    /// Mirrors Haskell's per-connection timeout draw (#701).
-    must_reply_timeout: Duration,
+    core: ServeCore,
 }
 
 impl ChainSyncServer {
@@ -173,7 +184,9 @@ impl ChainSyncServer {
     /// `[601 s, 911 s]` to match Haskell's `ouroboros-network` ChainSync
     /// codec timeout policy.
     pub fn new() -> Self {
-        Self::new_with_timeout(Self::draw_must_reply_timeout())
+        Self {
+            core: ServeCore::new(n2n_payload_encoder, "ChainSync"),
+        }
     }
 
     /// Test/explicit-timeout constructor.  Production code should call
@@ -182,28 +195,13 @@ impl ChainSyncServer {
     #[doc(hidden)]
     pub fn new_with_timeout(must_reply_timeout: Duration) -> Self {
         Self {
-            cursor_slot: 0,
-            cursor_hash: [0; 32],
-            cursor_initialized: false,
-            cursor_at_origin: false,
-            must_reply_timeout,
+            core: ServeCore::new_with_timeout(must_reply_timeout, n2n_payload_encoder, "ChainSync"),
         }
-    }
-
-    /// Draw a fresh `StMustReply` timeout uniformly in
-    /// `[CHAINSYNC_MUST_REPLY_TIMEOUT_MIN, CHAINSYNC_MUST_REPLY_TIMEOUT_MAX]`.
-    fn draw_must_reply_timeout() -> Duration {
-        use rand::Rng;
-        let span = CHAINSYNC_MUST_REPLY_TIMEOUT_MAX - CHAINSYNC_MUST_REPLY_TIMEOUT_MIN;
-        let span_ms = span.as_millis() as u64;
-        let mut rng = rand::rng();
-        let offset_ms = rng.random_range(0..=span_ms);
-        CHAINSYNC_MUST_REPLY_TIMEOUT_MIN + Duration::from_millis(offset_ms)
     }
 
     /// Configured `StMustReply` timeout (exposed for diagnostics / tests).
     pub fn must_reply_timeout(&self) -> Duration {
-        self.must_reply_timeout
+        self.core.must_reply_timeout()
     }
 
     /// Run the ChainSync server loop.
@@ -216,495 +214,23 @@ impl ChainSyncServer {
         &mut self,
         channel: &mut MuxChannel,
         block_provider: &B,
-        mut announcement_rx: broadcast::Receiver<BlockAnnouncement>,
-        mut rollback_rx: broadcast::Receiver<RollbackAnnouncement>,
+        announcement_rx: broadcast::Receiver<BlockAnnouncement>,
+        rollback_rx: broadcast::Receiver<RollbackAnnouncement>,
     ) -> Result<(), ProtocolError> {
-        let task_id = self as *const _ as usize;
-        tracing::info!(task_id, "chainsync server: task started");
-        loop {
-            tracing::debug!(
-                task_id,
-                cursor_slot = self.cursor_slot,
-                cursor_at_origin = self.cursor_at_origin,
-                "chainsync server: awaiting next client message"
-            );
-            let msg_bytes = channel.recv().await.map_err(|e| {
-                tracing::info!(
-                    task_id,
-                    error = %e,
-                    cursor_slot = self.cursor_slot,
-                    "chainsync server: channel.recv() failed — task exiting"
-                );
-                ProtocolError::from(e)
-            })?;
-            let msg = decode_message(&msg_bytes).map_err(|e| ProtocolError::CborDecode {
-                protocol: "ChainSync",
-                reason: e,
-            })?;
-
-            match msg {
-                ChainSyncMessage::MsgFindIntersect(points) => {
-                    self.handle_find_intersect(channel, block_provider, &points)
-                        .await?;
-                }
-                ChainSyncMessage::MsgRequestNext => {
-                    self.handle_request_next(
-                        channel,
-                        block_provider,
-                        &mut announcement_rx,
-                        &mut rollback_rx,
-                    )
-                    .await?;
-                }
-                ChainSyncMessage::MsgDone => {
-                    tracing::debug!("chainsync server: client sent MsgDone");
-                    return Ok(());
-                }
-                other => {
-                    return Err(ProtocolError::AgencyViolation {
-                        protocol: "ChainSync",
-                        state: "StIdle".to_string(),
-                        received_tag: format!("{other:?}")
-                            .as_bytes()
-                            .first()
-                            .copied()
-                            .unwrap_or(0),
-                    });
-                }
-            }
-        }
-    }
-
-    /// Handle MsgFindIntersect: walk the client's points to find the best match.
-    async fn handle_find_intersect<B: BlockProvider>(
-        &mut self,
-        channel: &mut MuxChannel,
-        block_provider: &B,
-        points: &[Point],
-    ) -> Result<(), ProtocolError> {
-        let tip = block_provider.get_tip();
-
-        // Walk points in order (most recent first), find the first one we have
-        for point in points {
-            match point {
-                Point::Origin => {
-                    // We always have genesis
-                    self.cursor_slot = 0;
-                    self.cursor_hash = [0; 32];
-                    self.cursor_initialized = true;
-                    self.cursor_at_origin = true;
-
-                    let response = encode_message(&ChainSyncMessage::MsgIntersectFound {
-                        point: Point::Origin,
-                        tip_slot: tip.slot,
-                        tip_hash: tip.hash,
-                        tip_block_number: tip.block_number,
-                    });
-                    channel.send(response).await.map_err(ProtocolError::from)?;
-                    return Ok(());
-                }
-                Point::Specific(slot, hash) => {
-                    if block_provider.has_block(hash) {
-                        self.cursor_slot = *slot;
-                        self.cursor_hash = *hash;
-                        self.cursor_initialized = true;
-                        self.cursor_at_origin = false;
-
-                        let response = encode_message(&ChainSyncMessage::MsgIntersectFound {
-                            point: point.clone(),
-                            tip_slot: tip.slot,
-                            tip_hash: tip.hash,
-                            tip_block_number: tip.block_number,
-                        });
-                        channel.send(response).await.map_err(ProtocolError::from)?;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        // No intersection found
-        let response = encode_message(&ChainSyncMessage::MsgIntersectNotFound {
-            tip_slot: tip.slot,
-            tip_hash: tip.hash,
-            tip_block_number: tip.block_number,
-        });
-        channel.send(response).await.map_err(ProtocolError::from)?;
-        Ok(())
-    }
-
-    /// Handle MsgRequestNext: serve the next header, propagate rollback, or wait
-    /// for an announcement.
-    ///
-    /// When a rollback is detected (either because the cursor points beyond the
-    /// rollback point, or a `RollbackAnnouncement` arrives while waiting at the
-    /// tip), the server sends `MsgRollBackward` to rewind the downstream peer's
-    /// cursor.
-    async fn handle_request_next<B: BlockProvider>(
-        &mut self,
-        channel: &mut MuxChannel,
-        block_provider: &B,
-        announcement_rx: &mut broadcast::Receiver<BlockAnnouncement>,
-        rollback_rx: &mut broadcast::Receiver<RollbackAnnouncement>,
-    ) -> Result<(), ProtocolError> {
-        // ── Check for pending rollbacks before serving ──────────────────────
-        // Drain any rollback announcements that arrived between the last
-        // MsgRequestNext and now.  If the cursor is beyond the rollback point,
-        // we must send MsgRollBackward instead of the next block.
-        if let Some(rb) = Self::drain_rollback(rollback_rx) {
-            if self.cursor_slot > rb.slot
-                || (self.cursor_slot == rb.slot && self.cursor_hash != rb.hash)
-            {
-                return self.send_rollback(channel, block_provider, &rb).await;
-            }
-        }
-
-        // ── Drain buffered block announcements ──────────────────────────────
-        // Block announcements are broadcast as dugite forges blocks, and are
-        // buffered in the receiver.  The direct `next_block` lookup below is
-        // authoritative — dugite commits each block to ChainDB before
-        // broadcasting the announcement, so any block we need to serve is
-        // already reachable via the block provider.  Drain all buffered
-        // announcements here so that the MsgAwaitReply loop only wakes on
-        // genuinely fresh post-drain forging events; otherwise a stale
-        // announcement for an already-served block would be re-served and
-        // rejected by the downstream peer with `UnexpectedBlockNo`.
-        loop {
-            match announcement_rx.try_recv() {
-                Ok(_) => continue,
-                Err(broadcast::error::TryRecvError::Empty) => break,
-                Err(broadcast::error::TryRecvError::Lagged(_)) => continue,
-                Err(broadcast::error::TryRecvError::Closed) => break,
-            }
-        }
-
-        // Try to find and serve the next block after our cursor.
-        let task_id = self as *const _ as usize;
-        let pre_lookup_cursor = self.cursor_slot;
-        if self.try_serve_next_block(channel, block_provider).await? {
-            tracing::debug!(
-                task_id,
-                pre_lookup_cursor,
-                post_serve_cursor = self.cursor_slot,
-                "chainsync server: direct serve path produced a message"
-            );
-            return Ok(());
-        }
-
-        // We're at the tip — send MsgAwaitReply and wait for announcement.
-        tracing::info!(
-            task_id,
-            cursor_slot = self.cursor_slot,
-            "chainsync server: no next block — sending MsgAwaitReply, entering StMustReply"
-        );
-        let await_msg = encode_message(&ChainSyncMessage::MsgAwaitReply);
-        channel.send(await_msg).await.map_err(|e| {
-            tracing::warn!(
-                task_id,
-                error = %e,
-                "chainsync server: MsgAwaitReply send failed"
-            );
-            ProtocolError::from(e)
-        })?;
-
-        // Wait for a block announcement OR rollback.  Timeout is the
-        // per-connection random draw in [601 s, 911 s] — matches Haskell's
-        // `minChainSyncTimeout`/`maxChainSyncTimeout` for non-trustable peers
-        // and prevents re-poll synchronization across a population of peers.
-        // Issue #701.
-        let timeout = self.must_reply_timeout;
-
-        // Bug J fix (2026-05-16): biased select so rollback_rx is always
-        // checked first when both arms are ready.  Combined with the cursor
-        // revalidation in `try_serve_next_block`, this eliminates the race
-        // where an announcement arriving simultaneously with a rollback
-        // would advance the cursor past the new chain's earlier blocks.
-        tokio::select! {
-            biased;
-
-            // ── Rollback received while waiting at tip ──────────────────────
-            // This is the critical path for issue #299: a follower in
-            // MsgAwaitReply (StMustReply) must receive MsgRollBackward when
-            // the chain switches to a better fork.
-            rollback = rollback_rx.recv() => {
-                match rollback {
-                    Ok(rb) => {
-                        // Only send rollback if cursor is at or beyond the rollback point.
-                        // If cursor is behind, the peer hasn't seen the rolled-back blocks
-                        // yet and will naturally follow the new fork.
-                        if self.cursor_slot > rb.slot
-                            || (self.cursor_slot == rb.slot && self.cursor_hash != rb.hash)
-                        {
-                            self.send_rollback(channel, block_provider, &rb).await
-                        } else {
-                            // Cursor is behind the rollback point — route through
-                            // `try_serve_next_block` so cursor revalidation (Bug J)
-                            // still applies on the post-rollback chain.
-                            self.try_serve_next_block(channel, block_provider).await?;
-                            Ok(())
-                        }
-                    }
-                    Err(broadcast::error::RecvError::Lagged(_)) => {
-                        // Missed rollback events — send rollback to current tip.
-                        let tip = block_provider.get_tip();
-                        let rb = RollbackAnnouncement {
-                            slot: tip.slot,
-                            hash: tip.hash,
-                        };
-                        self.send_rollback(channel, block_provider, &rb).await
-                    }
-                    Err(broadcast::error::RecvError::Closed) => Ok(()),
-                }
-            }
-            announcement = announcement_rx.recv() => {
-                tracing::info!(
-                    task_id,
-                    cursor_slot = self.cursor_slot,
-                    "chainsync server: woke from StMustReply on announcement"
-                );
-                match announcement {
-                    Ok(_ann) => {
-                        // The announcement is used only as a wake signal — the
-                        // block_provider is authoritative, and we MUST re-use
-                        // the same cursor-aware lookup as the direct-serve path
-                        // so that (a) slot-0 blocks at Origin are handled via
-                        // the inclusive `>=` lookup and (b) `cursor_at_origin`
-                        // is reliably cleared after the first serve.  Trusting
-                        // `ann.hash`/`ann.slot` directly skips the `cursor_at_origin
-                        // = false` transition and causes the next
-                        // MsgRequestNext to re-serve the first block, which
-                        // Haskell rejects with `UnexpectedBlockNo`.
-                        //
-                        // Bug J fix (2026-05-16): defence-in-depth — drain any
-                        // queued rollback first so the announcement-first race
-                        // is closed even if `try_serve_next_block`'s cursor
-                        // revalidation somehow missed (e.g., the cursor was
-                        // still on the old chain but the rollback queue had a
-                        // pending event).
-                        if let Some(rb) = Self::drain_rollback(rollback_rx) {
-                            if self.cursor_slot > rb.slot
-                                || (self.cursor_slot == rb.slot && self.cursor_hash != rb.hash)
-                            {
-                                return self.send_rollback(channel, block_provider, &rb).await;
-                            }
-                        }
-                        self.try_serve_next_block(channel, block_provider).await?;
-                        Ok(())
-                    }
-                    Err(broadcast::error::RecvError::Lagged(n)) => {
-                        tracing::warn!(
-                            n,
-                            "chainsync server: announcement channel lagged; catching up from tip"
-                        );
-                        // Route through `try_serve_next_block` so cursor
-                        // revalidation still applies after the lag.
-                        self.try_serve_next_block(channel, block_provider).await?;
-                        Ok(())
-                    }
-                    Err(broadcast::error::RecvError::Closed) => {
-                        // Broadcast sender dropped — node is shutting down.
-                        Ok(())
-                    }
-                }
-            }
-            _ = tokio::time::sleep(timeout) => {
-                tracing::warn!(
-                    task_id,
-                    cursor_slot = self.cursor_slot,
-                    timeout_seconds = timeout.as_secs(),
-                    "chainsync server: StMustReply timed out — re-polling block provider for next block"
-                );
-                self.try_serve_next_block(channel, block_provider).await?;
-                Ok(())
-            }
-        }
-    }
-
-    /// Look up the next block after the cursor and send it as `MsgRollForward`.
-    ///
-    /// This is the single authoritative serve path used by every wake-up site
-    /// in `handle_request_next` (direct path, announcement wake-up, timeout
-    /// poll). It respects `cursor_at_origin` by using the inclusive `>=`
-    /// lookup so a block at slot 0 (e.g. Byron genesis EBB) is not skipped,
-    /// and it unconditionally clears `cursor_at_origin` after a successful
-    /// serve.
-    ///
-    /// # Cursor revalidation (Bug J fix, 2026-05-16)
-    ///
-    /// Before serving, validate that the cursor's block is still on the
-    /// canonical chain.  If a fork switch has displaced it, send
-    /// `MsgRollBackward` to the most recent ancestor still on chain
-    /// instead of `MsgRollForward`.  This matches the Haskell ChainSync
-    /// server's behaviour: the follower's cursor anchor must always be
-    /// on the chain that the server is currently serving.
-    ///
-    /// Without this check, `get_next_block_after_slot(cursor_slot)`
-    /// silently skips any new-chain blocks whose slots fall ≤ `cursor_slot`
-    /// after a chain switch — the empirical Bug J failure mode.
-    ///
-    /// Returns `Ok(true)` if a message was sent (either `MsgRollForward` or
-    /// `MsgRollBackward`), `Ok(false)` if there is no next block to serve
-    /// (caller should continue waiting at the tip).
-    async fn try_serve_next_block<B: BlockProvider>(
-        &mut self,
-        channel: &mut MuxChannel,
-        block_provider: &B,
-    ) -> Result<bool, ProtocolError> {
-        let task_id = self as *const _ as usize;
-        // ── Cursor revalidation (Bug J) ─────────────────────────────────────
-        // Skip when the cursor is at Origin — the all-zero hash is by design
-        // not a stored block, so `is_on_chain` would always be false and we'd
-        // emit a redundant Origin → Origin rollback.
-        if !self.cursor_at_origin && !block_provider.is_on_chain(&self.cursor_hash) {
-            let rb = match block_provider.find_chain_ancestor(&self.cursor_hash) {
-                Some((slot, hash, _bn)) => RollbackAnnouncement { slot, hash },
-                // No ancestor reachable on chain — rewind all the way to
-                // Origin and let the peer resync from genesis.
-                None => RollbackAnnouncement {
-                    slot: 0,
-                    hash: [0u8; 32],
-                },
-            };
-            tracing::info!(
-                task_id,
-                cursor_slot = self.cursor_slot,
-                rewind_slot = rb.slot,
-                "chainsync server: cursor rolled off chain; \
-                 sending MsgRollBackward to ancestor"
-            );
-            self.send_rollback(channel, block_provider, &rb).await?;
-            return Ok(true);
-        }
-
-        // When cursor_at_origin is true, use the inclusive `>=` lookup so that
-        // a block at slot 0 is not skipped.  The strict `>` lookup would miss
-        // it since cursor_slot is 0.
-        //
-        // Otherwise advance by POINT, not by slot: a Byron EBB shares its
-        // absolute slot with the first main block of the epoch, and a
-        // slot-only advance from the EBB would skip that main block,
-        // serving the peer a chain with a hole in it.
-        let lookup_start = std::time::Instant::now();
-        let next_block = if self.cursor_at_origin {
-            block_provider.get_block_at_or_after_slot(0)
-        } else {
-            block_provider.get_next_block_after_point(self.cursor_slot, &self.cursor_hash)
-        };
-        let lookup_elapsed = lookup_start.elapsed();
-        if lookup_elapsed.as_millis() > 50 {
-            tracing::warn!(
-                task_id,
-                cursor_slot = self.cursor_slot,
-                elapsed_ms = lookup_elapsed.as_millis() as u64,
-                "chainsync server: block_provider lookup slow"
-            );
-        }
-
-        let Some((slot, hash, block_cbor)) = next_block else {
-            return Ok(false);
-        };
-
-        // Extract the block header and encode it as the HFC-wrapped header
-        // payload expected by N2N ChainSync: [era_id, #6.24(bstr(header_cbor))].
-        let hfc_header = extract_header_for_chainsync(&block_cbor).map_err(|reason| {
-            ProtocolError::CborDecode {
-                protocol: "ChainSync",
-                reason: format!("header extraction failed for block at slot {slot}: {reason}"),
-            }
-        })?;
-
-        let tip = block_provider.get_tip();
-        let response = encode_message(&ChainSyncMessage::MsgRollForward {
-            header: hfc_header,
-            tip_slot: tip.slot,
-            tip_hash: tip.hash,
-            tip_block_number: tip.block_number,
-        });
-        let send_start = std::time::Instant::now();
-        let send_result = channel.send(response).await;
-        let send_elapsed = send_start.elapsed();
-        if send_elapsed.as_millis() > 100 {
-            tracing::warn!(
-                task_id,
-                cursor_slot = self.cursor_slot,
-                serve_slot = slot,
-                send_elapsed_ms = send_elapsed.as_millis() as u64,
-                ok = send_result.is_ok(),
-                "chainsync server: MsgRollForward send slow"
-            );
-        }
-        send_result.map_err(ProtocolError::from)?;
-        tracing::info!(
-            task_id,
-            cursor_slot = self.cursor_slot,
-            serve_slot = slot,
-            "chainsync server: served MsgRollForward"
-        );
-
-        // Advance cursor and — critically — clear cursor_at_origin so the
-        // next MsgRequestNext uses the strict `>` lookup and does not
-        // re-serve the same block.
-        self.cursor_slot = slot;
-        self.cursor_hash = hash;
-        self.cursor_at_origin = false;
-        Ok(true)
+        self.core
+            .run(channel, block_provider, announcement_rx, rollback_rx)
+            .await
     }
 
     /// Drain any pending rollback announcements, returning the most recent one.
     ///
     /// Multiple rollbacks may have queued up between calls — only the latest
     /// matters because each successive rollback supersedes the previous one.
-    fn drain_rollback(
+    #[doc(hidden)]
+    pub fn drain_rollback(
         rollback_rx: &mut broadcast::Receiver<RollbackAnnouncement>,
     ) -> Option<RollbackAnnouncement> {
-        let mut latest: Option<RollbackAnnouncement> = None;
-        loop {
-            match rollback_rx.try_recv() {
-                Ok(rb) => latest = Some(rb),
-                Err(broadcast::error::TryRecvError::Lagged(_)) => {
-                    // Missed some — continue draining to get latest.
-                    continue;
-                }
-                Err(_) => break,
-            }
-        }
-        latest
-    }
-
-    /// Send `MsgRollBackward` to the downstream peer and rewind the cursor.
-    async fn send_rollback<B: BlockProvider>(
-        &mut self,
-        channel: &mut MuxChannel,
-        block_provider: &B,
-        rb: &RollbackAnnouncement,
-    ) -> Result<(), ProtocolError> {
-        let tip = block_provider.get_tip();
-        let point = if rb.slot == 0 && rb.hash == [0u8; 32] {
-            Point::Origin
-        } else {
-            Point::Specific(rb.slot, rb.hash)
-        };
-
-        tracing::info!(
-            rollback_slot = rb.slot,
-            cursor_slot = self.cursor_slot,
-            "chainsync server: sending MsgRollBackward to downstream peer"
-        );
-
-        let response = encode_message(&ChainSyncMessage::MsgRollBackward {
-            point: point.clone(),
-            tip_slot: tip.slot,
-            tip_hash: tip.hash,
-            tip_block_number: tip.block_number,
-        });
-        channel.send(response).await.map_err(ProtocolError::from)?;
-
-        // Rewind cursor to the rollback point.
-        self.cursor_slot = rb.slot;
-        self.cursor_hash = rb.hash;
-        self.cursor_at_origin = matches!(point, Point::Origin);
-
-        Ok(())
+        ServeCore::drain_rollback(rollback_rx)
     }
 }
 
@@ -716,8 +242,10 @@ impl Default for ChainSyncServer {
 
 #[cfg(test)]
 mod tests {
+    use super::super::serve_core::test_support::{
+        ForkAwareMockProvider, MockBlockProvider, MutableMockBlockProvider,
+    };
     use super::*;
-    use crate::TipInfo;
     use bytes::Bytes;
     use minicbor::Encoder;
     use tokio::sync::mpsc;
@@ -779,62 +307,6 @@ mod tests {
         // (this is what tag24 wraps — the CBOR serialisation of the header).
         enc.bytes(header_cbor).unwrap();
         buf
-    }
-
-    /// Mock block provider that stores (slot, hash, block_cbor) triples.
-    struct MockBlockProvider {
-        blocks: Vec<(u64, [u8; 32], Vec<u8>)>,
-    }
-
-    impl BlockProvider for MockBlockProvider {
-        fn get_block(&self, hash: &[u8; 32]) -> Option<Vec<u8>> {
-            self.blocks
-                .iter()
-                .find(|(_, h, _)| h == hash)
-                .map(|(_, _, cbor)| cbor.clone())
-        }
-
-        fn has_block(&self, hash: &[u8; 32]) -> bool {
-            self.blocks.iter().any(|(_, h, _)| h == hash)
-        }
-
-        fn get_tip(&self) -> TipInfo {
-            if let Some((slot, hash, _)) = self.blocks.last() {
-                TipInfo {
-                    slot: *slot,
-                    hash: *hash,
-                    block_number: self.blocks.len() as u64,
-                }
-            } else {
-                TipInfo {
-                    slot: 0,
-                    hash: [0; 32],
-                    block_number: 0,
-                }
-            }
-        }
-
-        fn get_next_block_after_slot(&self, after_slot: u64) -> Option<(u64, [u8; 32], Vec<u8>)> {
-            self.blocks
-                .iter()
-                .find(|(s, _, _)| *s > after_slot)
-                .cloned()
-        }
-
-        fn get_next_block_after_point(
-            &self,
-            slot: u64,
-            hash: &[u8; 32],
-        ) -> Option<(u64, [u8; 32], Vec<u8>)> {
-            if let Some(pos) = self
-                .blocks
-                .iter()
-                .position(|(s, h, _)| *s == slot && h == hash)
-            {
-                return self.blocks.get(pos + 1).cloned();
-            }
-            self.get_next_block_after_slot(slot)
-        }
     }
 
     fn make_test_channel() -> (
@@ -1141,71 +613,6 @@ mod tests {
 
     // ─── rollback propagation tests ─────────────────────────────────────────────
 
-    /// Block slot, hash, and CBOR data — the elements stored in mock providers.
-    type BlockEntry = (u64, [u8; 32], Vec<u8>);
-
-    /// Mock block provider that supports mutation for simulating rollback scenarios.
-    ///
-    /// Wraps blocks in an `Arc<Mutex<_>>` so both the server task and the test
-    /// can modify the block set concurrently (simulating ChainDB changes during
-    /// a fork switch).
-    struct MutableMockBlockProvider {
-        blocks: std::sync::Arc<std::sync::Mutex<Vec<BlockEntry>>>,
-    }
-
-    impl MutableMockBlockProvider {
-        fn new(blocks: Vec<BlockEntry>) -> Self {
-            Self {
-                blocks: std::sync::Arc::new(std::sync::Mutex::new(blocks)),
-            }
-        }
-    }
-
-    impl BlockProvider for MutableMockBlockProvider {
-        fn get_block(&self, hash: &[u8; 32]) -> Option<Vec<u8>> {
-            self.blocks
-                .lock()
-                .unwrap()
-                .iter()
-                .find(|(_, h, _)| h == hash)
-                .map(|(_, _, cbor)| cbor.clone())
-        }
-
-        fn has_block(&self, hash: &[u8; 32]) -> bool {
-            self.blocks
-                .lock()
-                .unwrap()
-                .iter()
-                .any(|(_, h, _)| h == hash)
-        }
-
-        fn get_tip(&self) -> TipInfo {
-            let blocks = self.blocks.lock().unwrap();
-            if let Some((slot, hash, _)) = blocks.last() {
-                TipInfo {
-                    slot: *slot,
-                    hash: *hash,
-                    block_number: blocks.len() as u64,
-                }
-            } else {
-                TipInfo {
-                    slot: 0,
-                    hash: [0; 32],
-                    block_number: 0,
-                }
-            }
-        }
-
-        fn get_next_block_after_slot(&self, after_slot: u64) -> Option<(u64, [u8; 32], Vec<u8>)> {
-            self.blocks
-                .lock()
-                .unwrap()
-                .iter()
-                .find(|(s, _, _)| *s > after_slot)
-                .cloned()
-        }
-    }
-
     #[tokio::test]
     async fn rollback_sends_msg_roll_backward() {
         // Scenario: serve 3 blocks (slots 10, 20, 30), then trigger a 2-block
@@ -1437,153 +844,6 @@ mod tests {
         let done = encode_message(&ChainSyncMessage::MsgDone);
         ingress_tx.send(Bytes::from(done)).await.unwrap();
         handle.await.unwrap().unwrap();
-    }
-
-    // ─── Bug J: cursor revalidation after fork switch (2026-05-16) ───────────
-
-    /// Fork-aware mock that distinguishes blocks on the *canonical chain*
-    /// from blocks merely stored as forks.  This is required to exercise
-    /// the Bug J fix in the ChainSync server: `try_serve_next_block` must
-    /// detect when the cursor's block has been displaced by a fork switch
-    /// (off-chain) and send `MsgRollBackward` rather than skipping over
-    /// the new chain's earlier-slot blocks.
-    type StoredBlock = (u64, u64, [u8; 32], Vec<u8>); // (slot, block_no, prev_hash, cbor)
-
-    struct ForkAwareMockProvider {
-        /// Hashes on the current canonical chain, oldest → newest.
-        chain: std::sync::Arc<std::sync::Mutex<Vec<[u8; 32]>>>,
-        /// Every stored block keyed by hash.
-        store: std::sync::Arc<std::sync::Mutex<std::collections::HashMap<[u8; 32], StoredBlock>>>,
-    }
-
-    impl ForkAwareMockProvider {
-        fn new() -> Self {
-            Self {
-                chain: std::sync::Arc::new(std::sync::Mutex::new(Vec::new())),
-                store: std::sync::Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
-            }
-        }
-
-        /// Append a block to the canonical chain.
-        fn push_on_chain(
-            &self,
-            slot: u64,
-            hash: [u8; 32],
-            prev_hash: [u8; 32],
-            block_no: u64,
-            cbor: Vec<u8>,
-        ) {
-            self.store
-                .lock()
-                .unwrap()
-                .insert(hash, (slot, block_no, prev_hash, cbor));
-            self.chain.lock().unwrap().push(hash);
-        }
-
-        /// Store a block as a fork (not on canonical chain).
-        fn put_fork(
-            &self,
-            slot: u64,
-            hash: [u8; 32],
-            prev_hash: [u8; 32],
-            block_no: u64,
-            cbor: Vec<u8>,
-        ) {
-            self.store
-                .lock()
-                .unwrap()
-                .insert(hash, (slot, block_no, prev_hash, cbor));
-        }
-
-        /// Replace the canonical chain wholesale (simulates a fork switch).
-        /// All previously-on-chain blocks not in `new_chain` become forks.
-        fn replace_chain(&self, new_chain: Vec<[u8; 32]>) {
-            *self.chain.lock().unwrap() = new_chain;
-        }
-    }
-
-    impl BlockProvider for ForkAwareMockProvider {
-        fn get_block(&self, hash: &[u8; 32]) -> Option<Vec<u8>> {
-            self.store
-                .lock()
-                .unwrap()
-                .get(hash)
-                .map(|(_, _, _, cbor)| cbor.clone())
-        }
-
-        fn has_block(&self, hash: &[u8; 32]) -> bool {
-            self.store.lock().unwrap().contains_key(hash)
-        }
-
-        fn get_tip(&self) -> TipInfo {
-            let chain = self.chain.lock().unwrap();
-            let store = self.store.lock().unwrap();
-            if let Some(tip_hash) = chain.last() {
-                if let Some((slot, block_no, _, _)) = store.get(tip_hash) {
-                    return TipInfo {
-                        slot: *slot,
-                        hash: *tip_hash,
-                        block_number: *block_no,
-                    };
-                }
-            }
-            TipInfo {
-                slot: 0,
-                hash: [0; 32],
-                block_number: 0,
-            }
-        }
-
-        fn get_next_block_after_slot(&self, after_slot: u64) -> Option<(u64, [u8; 32], Vec<u8>)> {
-            // Walk the canonical chain in order looking for the first block
-            // with slot > after_slot.  Mirrors VolatileDB's behaviour.
-            let chain = self.chain.lock().unwrap();
-            let store = self.store.lock().unwrap();
-            for hash in chain.iter() {
-                if let Some((slot, _, _, cbor)) = store.get(hash) {
-                    if *slot > after_slot {
-                        return Some((*slot, *hash, cbor.clone()));
-                    }
-                }
-            }
-            None
-        }
-
-        fn get_block_at_or_after_slot(&self, slot: u64) -> Option<(u64, [u8; 32], Vec<u8>)> {
-            let chain = self.chain.lock().unwrap();
-            let store = self.store.lock().unwrap();
-            for hash in chain.iter() {
-                if let Some((s, _, _, cbor)) = store.get(hash) {
-                    if *s >= slot {
-                        return Some((*s, *hash, cbor.clone()));
-                    }
-                }
-            }
-            None
-        }
-
-        fn is_on_chain(&self, hash: &[u8; 32]) -> bool {
-            self.chain.lock().unwrap().iter().any(|h| h == hash)
-        }
-
-        fn find_chain_ancestor(&self, start_hash: &[u8; 32]) -> Option<(u64, [u8; 32], u64)> {
-            let chain: std::collections::HashSet<[u8; 32]> =
-                self.chain.lock().unwrap().iter().copied().collect();
-            let store = self.store.lock().unwrap();
-            let mut current = *start_hash;
-            let mut visited = std::collections::HashSet::new();
-            loop {
-                if !visited.insert(current) {
-                    return None;
-                }
-                if chain.contains(&current) {
-                    let (slot, block_no, _, _) = store.get(&current)?;
-                    return Some((*slot, current, *block_no));
-                }
-                let (_, _, prev_hash, _) = store.get(&current)?;
-                current = *prev_hash;
-            }
-        }
     }
 
     /// **Bug J regression** (issue #500, 2026-05-16):
@@ -2348,6 +1608,249 @@ mod tests {
         }
 
         // Clean up.
+        let done = encode_message(&ChainSyncMessage::MsgDone);
+        ingress_tx.send(Bytes::from(done)).await.unwrap();
+        handle.await.unwrap().unwrap();
+    }
+
+    // ─── #869: StMustReply silent no-reply ───────────────────────────────────
+
+    /// Regression test for #869: a spurious announcement wake (nothing new
+    /// past the cursor) must NOT cause `handle_request_next` to return
+    /// without sending a message.
+    ///
+    /// Per the Ouroboros ChainSync state machine, once `MsgAwaitReply` has
+    /// been sent the client is in `StMustReply` and will not send another
+    /// `MsgRequestNext` until it receives a reply.  Pre-fix, a spurious wake
+    /// (e.g. an announcement whose content doesn't correspond to anything
+    /// new past the cursor) caused the server to silently return `Ok(())`
+    /// without sending anything — the outer loop would then block on
+    /// `channel.recv()` waiting for a client message that, per protocol,
+    /// will never arrive.  This test drives that exact sequence without ever
+    /// sending a second `MsgRequestNext`, proving the server keeps waiting
+    /// on the SAME `handle_request_next` call until it has something to send.
+    #[tokio::test]
+    async fn spurious_announcement_wake_does_not_return_agency_without_reply() {
+        let block_a = make_hfc_block(7, &[0x0A]);
+        let provider = MutableMockBlockProvider::new(vec![(10, [0x01; 32], block_a)]);
+        let blocks_ref = provider.blocks.clone();
+
+        let (mut channel, mut egress_rx, ingress_tx) = make_test_channel();
+        let (ann_tx, _) = broadcast::channel::<BlockAnnouncement>(16);
+        let (rb_tx, _) = broadcast::channel::<RollbackAnnouncement>(16);
+        let ann_rx = ann_tx.subscribe();
+        let rb_rx = rb_tx.subscribe();
+
+        let mut server = ChainSyncServer::new();
+        let handle =
+            tokio::spawn(async move { server.run(&mut channel, &provider, ann_rx, rb_rx).await });
+
+        // Intersect at origin, serve the only block — cursor now at slot 10.
+        let find = encode_message(&ChainSyncMessage::MsgFindIntersect(vec![Point::Origin]));
+        ingress_tx.send(Bytes::from(find)).await.unwrap();
+        let _ = egress_rx.recv().await.unwrap(); // MsgIntersectFound
+
+        let req = encode_message(&ChainSyncMessage::MsgRequestNext);
+        ingress_tx.send(Bytes::from(req)).await.unwrap();
+        let (_, _, resp) = egress_rx.recv().await.unwrap();
+        assert!(matches!(
+            decode_message(&resp).unwrap(),
+            ChainSyncMessage::MsgRollForward { .. }
+        ));
+
+        // Request next again — at tip, enters StMustReply.
+        let req = encode_message(&ChainSyncMessage::MsgRequestNext);
+        ingress_tx.send(Bytes::from(req)).await.unwrap();
+        let (_, _, resp) = egress_rx.recv().await.unwrap();
+        assert!(matches!(
+            decode_message(&resp).unwrap(),
+            ChainSyncMessage::MsgAwaitReply
+        ));
+
+        // Fire a SPURIOUS announcement: nothing new is added to the block
+        // store, so `try_serve_next_block` finds nothing to serve.
+        ann_tx
+            .send(BlockAnnouncement {
+                slot: 10,
+                hash: [0x01; 32],
+                block_number: 1,
+            })
+            .unwrap();
+
+        // Give the server a moment to process the spurious wake.  Pre-fix,
+        // this would cause `handle_request_next` to return without sending
+        // anything, wedging the connection.
+        assert!(
+            tokio::time::timeout(Duration::from_millis(200), egress_rx.recv())
+                .await
+                .is_err(),
+            "server must not send anything on a spurious wake with nothing new to serve"
+        );
+
+        // Now push a genuine new block and announce it — the SAME
+        // `handle_request_next` call (no second MsgRequestNext from the
+        // client) must eventually deliver it.
+        let block_b = make_hfc_block(7, &[0x0B]);
+        blocks_ref.lock().unwrap().push((20, [0x02; 32], block_b));
+        ann_tx
+            .send(BlockAnnouncement {
+                slot: 20,
+                hash: [0x02; 32],
+                block_number: 2,
+            })
+            .unwrap();
+
+        let (_, _, resp) = tokio::time::timeout(Duration::from_secs(5), egress_rx.recv())
+            .await
+            .expect("server must eventually reply once a real block is available")
+            .unwrap();
+        assert!(
+            matches!(
+                decode_message(&resp).unwrap(),
+                ChainSyncMessage::MsgRollForward { .. }
+            ),
+            "server must serve the genuine block without a second MsgRequestNext"
+        );
+
+        handle.abort();
+    }
+
+    // ─── #876: MsgFindIntersect fork/slot validation ─────────────────────────
+
+    /// #876: a client claiming the wrong slot for a real on-chain hash must
+    /// NOT get `MsgIntersectFound` — accepting it would poison the follower
+    /// cursor with a `(claimed_slot, real_hash)` pair that doesn't correspond
+    /// to any actual chain point (e.g. `(u64::MAX, real_hash)` would wedge
+    /// every subsequent `try_serve_next_block` lookup).
+    #[tokio::test]
+    async fn find_intersect_wrong_claimed_slot_rejected() {
+        let provider = ForkAwareMockProvider::new();
+        let genesis = [0u8; 32];
+        let a_hash = [0x0A; 32];
+        provider.push_on_chain(10, a_hash, genesis, 1, make_hfc_block(7, &[0xA1]));
+
+        let (mut channel, mut egress_rx, ingress_tx) = make_test_channel();
+        let (ann_tx, _) = broadcast::channel(16);
+        let (rb_tx, _) = broadcast::channel(16);
+        let mut server = ChainSyncServer::new();
+
+        let handle = tokio::spawn(async move {
+            server
+                .run(
+                    &mut channel,
+                    &provider,
+                    ann_tx.subscribe(),
+                    rb_tx.subscribe(),
+                )
+                .await
+        });
+
+        // a_hash is real and on-chain, but at slot 10 — not the claimed 999.
+        let find = encode_message(&ChainSyncMessage::MsgFindIntersect(vec![Point::Specific(
+            999, a_hash,
+        )]));
+        ingress_tx.send(Bytes::from(find)).await.unwrap();
+
+        let (_, _, resp) = egress_rx.recv().await.unwrap();
+        let msg = decode_message(&resp).unwrap();
+        assert!(
+            matches!(msg, ChainSyncMessage::MsgIntersectNotFound { .. }),
+            "wrong claimed slot for a real hash must be rejected, got {msg:?}"
+        );
+
+        let done = encode_message(&ChainSyncMessage::MsgDone);
+        ingress_tx.send(Bytes::from(done)).await.unwrap();
+        handle.await.unwrap().unwrap();
+    }
+
+    /// #876: a hash that exists in storage but is NOT on the canonical chain
+    /// (a fork block) must not be accepted as an intersection point — only
+    /// `is_on_chain` hashes may become the follower cursor.
+    #[tokio::test]
+    async fn find_intersect_fork_only_hash_rejected() {
+        let provider = ForkAwareMockProvider::new();
+        let genesis = [0u8; 32];
+        let a_hash = [0x0A; 32];
+        provider.push_on_chain(10, a_hash, genesis, 1, make_hfc_block(7, &[0xA1]));
+        let fork_hash = [0xF0; 32];
+        provider.put_fork(15, fork_hash, a_hash, 2, make_hfc_block(7, &[0xF1]));
+
+        let (mut channel, mut egress_rx, ingress_tx) = make_test_channel();
+        let (ann_tx, _) = broadcast::channel(16);
+        let (rb_tx, _) = broadcast::channel(16);
+        let mut server = ChainSyncServer::new();
+
+        let handle = tokio::spawn(async move {
+            server
+                .run(
+                    &mut channel,
+                    &provider,
+                    ann_tx.subscribe(),
+                    rb_tx.subscribe(),
+                )
+                .await
+        });
+
+        let find = encode_message(&ChainSyncMessage::MsgFindIntersect(vec![Point::Specific(
+            15, fork_hash,
+        )]));
+        ingress_tx.send(Bytes::from(find)).await.unwrap();
+
+        let (_, _, resp) = egress_rx.recv().await.unwrap();
+        let msg = decode_message(&resp).unwrap();
+        assert!(
+            matches!(msg, ChainSyncMessage::MsgIntersectNotFound { .. }),
+            "fork-only hash must not be accepted as an intersection point, got {msg:?}"
+        );
+
+        let done = encode_message(&ChainSyncMessage::MsgDone);
+        ingress_tx.send(Bytes::from(done)).await.unwrap();
+        handle.await.unwrap().unwrap();
+    }
+
+    /// #876: a correct canonical point (real on-chain hash + matching slot)
+    /// must still be accepted — the fix must not be overly strict.
+    #[tokio::test]
+    async fn find_intersect_correct_canonical_point_found() {
+        let provider = ForkAwareMockProvider::new();
+        let genesis = [0u8; 32];
+        let a_hash = [0x0A; 32];
+        let b_hash = [0x0B; 32];
+        provider.push_on_chain(10, a_hash, genesis, 1, make_hfc_block(7, &[0xA1]));
+        provider.push_on_chain(20, b_hash, a_hash, 2, make_hfc_block(7, &[0xB1]));
+
+        let (mut channel, mut egress_rx, ingress_tx) = make_test_channel();
+        let (ann_tx, _) = broadcast::channel(16);
+        let (rb_tx, _) = broadcast::channel(16);
+        let mut server = ChainSyncServer::new();
+
+        let handle = tokio::spawn(async move {
+            server
+                .run(
+                    &mut channel,
+                    &provider,
+                    ann_tx.subscribe(),
+                    rb_tx.subscribe(),
+                )
+                .await
+        });
+
+        let find = encode_message(&ChainSyncMessage::MsgFindIntersect(vec![Point::Specific(
+            10, a_hash,
+        )]));
+        ingress_tx.send(Bytes::from(find)).await.unwrap();
+
+        let (_, _, resp) = egress_rx.recv().await.unwrap();
+        let msg = decode_message(&resp).unwrap();
+        match msg {
+            ChainSyncMessage::MsgIntersectFound { point, .. } => {
+                assert_eq!(point, Point::Specific(10, a_hash));
+            }
+            other => {
+                panic!("expected MsgIntersectFound for correct canonical point, got {other:?}")
+            }
+        }
+
         let done = encode_message(&ChainSyncMessage::MsgDone);
         ingress_tx.send(Bytes::from(done)).await.unwrap();
         handle.await.unwrap().unwrap();
